pyproject.toml
[project]
name = "tts-ssml"
version = "0.1.0"
description = "Google Cloud Text-to-Speech APIを使用してSSMLファイルを音声に変換するツール"
requires-python = ">=3.8"
dependencies = [
    "google-cloud-texttospeech>=2.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [] 